+++ extensions-core/kafka-indexing-service/src/main/java/io/druid/indexing/kafka/KafkaIndexTask.java
@@ -24,13 +24,11 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
-import com.google.common.base.Supplier;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -44,7 +42,6 @@ import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.metamx.emitter.EmittingLogger;
 import io.druid.concurrent.Execs;
-import io.druid.data.input.Committer;
 import io.druid.data.input.InputRow;
 import io.druid.data.input.impl.InputRowParser;
 import io.druid.indexing.appenderator.ActionBasedSegmentAllocator;
@@ -53,12 +50,10 @@ import io.druid.indexing.common.TaskStatus;
 import io.druid.indexing.common.TaskToolbox;
 import io.druid.indexing.common.actions.CheckPointDataSourceMetadataAction;
 import io.druid.indexing.common.actions.ResetDataSourceMetadataAction;
-import io.druid.indexing.common.actions.SegmentTransactionalInsertAction;
 import io.druid.indexing.common.actions.TaskActionClient;
 import io.druid.indexing.common.task.AbstractTask;
 import io.druid.indexing.common.task.TaskResource;
 import io.druid.indexing.overlord.DataSourceMetadata;
-import io.druid.java.util.common.IAE;
 import io.druid.java.util.common.ISE;
 import io.druid.java.util.common.guava.Sequence;
 import io.druid.java.util.common.parsers.ParseException;
@@ -75,9 +70,7 @@ import io.druid.segment.realtime.RealtimeMetricsMonitor;
 import io.druid.segment.realtime.appenderator.Appenderator;
 import io.druid.segment.realtime.appenderator.Appenderators;
 import io.druid.segment.realtime.appenderator.FiniteAppenderatorDriver;
-import io.druid.segment.realtime.appenderator.SegmentIdentifier;
 import io.druid.segment.realtime.appenderator.SegmentsAndMetadata;
-import io.druid.segment.realtime.appenderator.TransactionalSegmentPublisher;
 import io.druid.segment.realtime.firehose.ChatHandler;
 import io.druid.segment.realtime.firehose.ChatHandlerProvider;
 import io.druid.timeline.DataSegment;
@@ -98,7 +91,6 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -142,7 +134,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
   private static final long POLL_TIMEOUT = 100;
   private static final long POLL_RETRY_MS = 30000;
   private static final long LOCK_ACQUIRE_TIMEOUT_SECONDS = 15;
-  private static final String METADATA_NEXT_PARTITIONS = "nextPartitions";
+  static final String METADATA_NEXT_PARTITIONS = "nextPartitions";
 
   private final DataSchema dataSchema;
   private final InputRowParser<ByteBuffer> parser;
@@ -203,7 +195,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
   private volatile boolean pauseRequested = false;
   private volatile long pauseMillis = 0;
 
-  private volatile int nextDriverIndex = 0;
+  volatile int nextDriverIndex = 0;
   // Reverse sorted list of DriverHolders i.e. the most recent driverHolder is at front and the oldest at last
   private final List<DriverHolder> driverHolders = new CopyOnWriteArrayList<>();
   private final Lock driversListLock = new ReentrantLock();
@@ -371,7 +363,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
                       log.error(
                           t,
                           "Error in publish thread, enqueueing driver [%d] back to publish queue",
-                          driverHolder.driverIndex
+                          driverHolder.getDriverIndex()
                       );
                       driverHolder.driverStatus = DriverHolder.DriverStatus.PERSISTED;
                       publishQueue.addFirst(driverHolder);
@@ -403,17 +395,17 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
                     driverHolder.driverStatus == DriverHolder.DriverStatus.PUBLISHED,
                     String.format(
                         "WTH?! cannot wait for hand off for not published driver [%d], status: [%s]",
-                        driverHolder.driverIndex,
+                        driverHolder.getDriverIndex(),
                         driverHolder.driverStatus
                     )
                 );
 
                 log.info("Waiting for driver [%s] to hand off", driverHolder.getMetadata());
                 try {
-                  if (driverHolder.driverIndex != -1) {
+                  if (driverHolder.getDriverIndex() != -1) {
                     driverHolder.waitForHandOff();
                   }
-                  log.info("Handoff complete for driver [%d]", driverHolder.driverIndex);
+                  log.info("Handoff complete for driver [%d]", driverHolder.getDriverIndex());
                 }
                 catch (InterruptedException t) {
                   if (stopRequested.get()) {
@@ -427,7 +419,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
                 finally {
                   driverHolder.close();
                   if (!driverHolders.remove(driverHolder)) {
-                    log.warn("Unable to remove driver [%d], it was not in the drivers list", driverHolder.driverIndex);
+                    log.warn("Unable to remove driver [%d], it was not in the drivers list", driverHolder.getDriverIndex());
                   } else {
                     try {
                       lockDriversList();
@@ -619,7 +611,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
                     lockDriversList();
                     final DriverHolder nextDriverHolder = DriverHolder.getNextDriverHolder(
                         this,
-                        driverHolders.get(0).endOffsets,
+                        driverHolders.get(0).getEndOffsets(),
                         maxEndOffsets,
                         ioConfig.getBaseSequenceName(),
                         fireDepartmentMetrics,
@@ -661,10 +653,10 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
             pause(-1L);
 
             KafkaDataSourceMetadata previousCheckPoint = null;
-            if (!latestDriver.startOffsets.equals(ioConfig.getStartPartitions().getPartitionOffsetMap())) {
+            if (!latestDriver.getStartOffsets().equals(ioConfig.getStartPartitions().getPartitionOffsetMap())) {
               previousCheckPoint = new KafkaDataSourceMetadata(new KafkaPartitions(
                   topic,
-                  latestDriver.startOffsets
+                  latestDriver.getStartOffsets()
               ));
             }
 
@@ -706,7 +698,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
           }
         }
         // add Sentinel Driver at the end so that task can wait till the persistLatch and handOffLatch is countdown
-        persistAndPossiblyPublish(new SentinelDriverHolder(persistLatch, handOffLatch));
+        persistAndPossiblyPublish(new DriverHolder.SentinelDriverHolder(persistLatch, handOffLatch));
       }
 
       persistLatch.await();
@@ -772,10 +764,10 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
             try {
               result = driverHolder.persist();
               driverHolder.driverStatus = DriverHolder.DriverStatus.PERSISTED;
-              log.info("Driver [%d] persisted with result [%s]", driverHolder.driverIndex, result);
+              log.info("Driver [%d] persisted with result [%s]", driverHolder.getDriverIndex(), result);
 
               if (stopRequested.get() && !publishOnStop.get()) {
-                log.warn("Skipping publish of driver [%d] as we are asked to stop", driverHolder.driverIndex);
+                log.warn("Skipping publish of driver [%d] as we are asked to stop", driverHolder.getDriverIndex());
               } else {
                 log.info("Adding driver to publish queue, [%s]", driverHolder.getMetadata());
                 publishQueue.addLast(driverHolder);
@@ -783,7 +775,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
             }
             catch (Exception e) {
               if (e instanceof InterruptedException && (stopRequested.get() || handOffLatch.getCount() == 0)) {
-                log.warn("Interrupted while persisting driver [%d], aborting persist", driverHolder.driverIndex);
+                log.warn("Interrupted while persisting driver [%d], aborting persist", driverHolder.getDriverIndex());
                 return null;
               }
               log.error("Error [%s] while persisting driver [%s]", e.getMessage(), driverHolder.getMetadata());
@@ -936,10 +928,10 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
     // save latest state on disk
     persistDriversList();
 
-    if (driverHolders.get(0).last) {
+    if (driverHolders.get(0).isLast()) {
       // recovered a driver which happens to be the last driver for this task
       // set the end offsets for this task so that the task shutdowns eventually
-      endOffsets.putAll(driverHolders.get(0).endOffsets);
+      endOffsets.putAll(driverHolders.get(0).getEndOffsets());
     }
   }
 
@@ -1024,7 +1016,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
                       @Override
                       public Sequence<T> run(Query<T> query, Map<String, Object> responseContext)
                       {
-                        return query.run(input.driver.getAppenderator(), responseContext);
+                        return query.run(input.getDriver().getAppenderator(), responseContext);
                       }
                     };
                   }
@@ -1115,27 +1107,27 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
       final DriverHolder driverHolder = driverHolders.get(0);
 
       try {
-        if (driverHolder.checkPointed) {
+        if (driverHolder.isCheckPointed()) {
           // this should only happen when we got another setEndOffsets call after the final setEndOffsets call
           // check for consistency and duplicate request
           Preconditions.checkState(
-              endOffsets.equals(driverHolder.endOffsets),
+              endOffsets.equals(driverHolder.getEndOffsets()),
               "WTH?! End offsets for task [%s] and latest driver [%s] do not match",
               endOffsets,
               driverHolder.getMetadata()
           );
           // ignore duplicate requests
-          if (offsets.equals(driverHolder.endOffsets)) {
+          if (offsets.equals(driverHolder.getEndOffsets())) {
             log.warn(
                 "end offsets already set to [%s], ignoring duplicate request to set to [%s]",
-                driverHolder.endOffsets,
+                driverHolder.getEndOffsets(),
                 offsets
             );
             return Response.ok(endOffsets).build();
           } else {
             throw new ISE(
                 "WTH?! end offsets set to [%s], ignoring request to set to [%s]",
-                driverHolder.endOffsets,
+                driverHolder.getEndOffsets(),
                 offsets
             );
           }
@@ -1144,13 +1136,13 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
 
         if (finish) {
           // set the last flag, useful while restoring state from disk to set endOffsets
-          driverHolder.last = true;
+          driverHolder.setLast();
           endOffsets.putAll(offsets);
         } else {
           // create next driver
           final DriverHolder nextDriverHolder = DriverHolder.getNextDriverHolder(
               this,
-              driverHolder.endOffsets, //previous driver endOffsets
+              driverHolder.getEndOffsets(), //previous driver endOffsets
               endOffsets, // task endOffsets
               ioConfig.getBaseSequenceName(),
               fireDepartmentMetrics,
@@ -1307,7 +1299,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
     return status == Status.PAUSED;
   }
 
-  private Appenderator newAppenderator(FireDepartmentMetrics metrics, TaskToolbox toolbox, File basePersistDirectory)
+  Appenderator newAppenderator(FireDepartmentMetrics metrics, TaskToolbox toolbox, File basePersistDirectory)
   {
     return Appenderators.createRealtime(
         dataSchema,
@@ -1326,7 +1318,7 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
     );
   }
 
-  private FiniteAppenderatorDriver newDriver(
+  FiniteAppenderatorDriver newDriver(
       final Appenderator appenderator,
       final TaskToolbox toolbox,
       final FireDepartmentMetrics metrics
@@ -1556,609 +1548,6 @@ public class KafkaIndexTask extends AbstractTask implements ChatHandler
       }
     } else {
       log.makeAlert("Failed to send reset request for partitions [%s]", partitionOffsetMap.keySet()).emit();
-    }
-  }
-
-  private static class DriverHolder implements Closeable
-  {
-    public enum DriverStatus
-    {
-      NOT_OPEN,
-      OPEN,
-      COMPLETE,
-      PERSISTING,
-      PERSISTED,
-      PUBLISHING,
-      PUBLISHED,
-      CLOSED
-    }
-
-    private volatile DriverStatus driverStatus = DriverStatus.NOT_OPEN;
-
-    private final FiniteAppenderatorDriver driver;
-    private final Set<Integer> assignment;
-    private Supplier<Committer> committerSupplier;
-
-    private final Map<Integer, Long> nextPartitionOffsets = Maps.newHashMap();
-
-    /*********** Driver Metadata **********/
-    private final String topic;
-    private final Map<Integer, Long> startOffsets;
-    private final Map<Integer, Long> endOffsets;
-    private final String sequenceName;
-    // this uniquely identifies the persist dir for this driver
-    private final int driverIndex;
-    // whether this is the last driver for the task
-    private volatile boolean last;
-    // whether the end offsets for this driver is set
-    private volatile boolean checkPointed;
-    private boolean maxRowsPerSegmentLimitReached;
-
-    /***************************************/
-
-    private DriverHolder(
-        String topic,
-        FiniteAppenderatorDriver driver,
-        Map<Integer, Long> startOffsets,
-        Map<Integer, Long> endOffsets,
-        int driverIndex,
-        String sequenceName,
-        boolean last,
-        boolean checkPointed,
-        boolean maxRowsPerSegmentLimitReached
-    )
-    {
-      this.topic = topic;
-      this.driver = driver;
-      this.startOffsets = startOffsets;
-      // endOffsets will change when a check point is set so make a local copy of it
-      this.endOffsets = Maps.newHashMap(endOffsets);
-      this.driverIndex = driverIndex;
-      this.sequenceName = sequenceName;
-      this.assignment = Sets.newHashSet(endOffsets.keySet());
-      this.last = last;
-      this.checkPointed = checkPointed;
-      this.maxRowsPerSegmentLimitReached = maxRowsPerSegmentLimitReached;
-    }
-
-    Map<Integer, Long> startJob(final ObjectMapper mapper)
-    {
-      final Object restored = driver.startJob();
-      Preconditions.checkState(
-          driverStatus == DriverStatus.NOT_OPEN,
-          "WTH?! Cannot change driver status to [%s] from [%s]",
-          DriverStatus.OPEN,
-          driverStatus
-      );
-      driverStatus = DriverStatus.OPEN;
-      if (restored == null) {
-        nextPartitionOffsets.putAll(startOffsets);
-      } else {
-        Map<String, Object> restoredMetadataMap = (Map) restored;
-        final KafkaPartitions restoredNextPartitions = mapper.convertValue(
-            restoredMetadataMap.get(METADATA_NEXT_PARTITIONS),
-            KafkaPartitions.class
-        );
-        nextPartitionOffsets.putAll(restoredNextPartitions.getPartitionOffsetMap());
-
-        // Sanity checks.
-        if (!restoredNextPartitions.getTopic().equals(topic)) {
-          throw new ISE(
-              "WTF?! Restored topic[%s] but expected topic[%s]",
-              restoredNextPartitions.getTopic(),
-              topic
-          );
-        }
-
-        if (!nextPartitionOffsets.keySet().equals(startOffsets.keySet())) {
-          throw new ISE(
-              "WTF?! Restored partitions[%s] but expected partitions[%s]",
-              nextPartitionOffsets.keySet(),
-              startOffsets.keySet()
-          );
-        }
-      }
-      committerSupplier = new Supplier<Committer>()
-      {
-        @Override
-        public Committer get()
-        {
-          final Map<Integer, Long> snapshot = ImmutableMap.copyOf(nextPartitionOffsets);
-
-          return new Committer()
-          {
-            @Override
-            public Object getMetadata()
-            {
-              return ImmutableMap.of(
-                  METADATA_NEXT_PARTITIONS, new KafkaPartitions(
-                      topic,
-                      snapshot
-                  )
-              );
-            }
-
-            @Override
-            public void run()
-            {
-              // Do nothing
-            }
-          };
-        }
-      };
-
-      // remove partition for which offsets have been consumed fully
-      for (Map.Entry<Integer, Long> partitionOffset : nextPartitionOffsets.entrySet()) {
-        if (endOffsets.get(partitionOffset.getKey()).equals(partitionOffset.getValue())) {
-          logInfo("[Start-Job] read partition [%d]", partitionOffset.getKey());
-          assignment.remove(partitionOffset.getKey());
-        }
-      }
-      if (assignment.isEmpty()) {
-        // finish the driver
-        logInfo("[Start-Job] read all partitions");
-        Preconditions.checkState(
-            driverStatus == DriverStatus.OPEN,
-            "WTH?! Cannot change driver status to [%s] from [%s]",
-            DriverStatus.COMPLETE,
-            driverStatus
-        );
-        driverStatus = DriverStatus.COMPLETE;
-      }
-
-      return nextPartitionOffsets;
-    }
-
-    boolean canHandle(ConsumerRecord<byte[], byte[]> record)
-    {
-      return driverStatus == DriverStatus.OPEN
-             && endOffsets.get(record.partition()) != null
-             && record.offset() >= startOffsets.get(record.partition())
-             && record.offset() < endOffsets.get(record.partition());
-    }
-
-    public SegmentIdentifier add(InputRow row) throws IOException
-    {
-      Preconditions.checkState(driverStatus == DriverStatus.OPEN, "Cannot add to driver which is not open!");
-
-      SegmentIdentifier identifier = driver.add(row, sequenceName, committerSupplier, false);
-
-      if (identifier == null) {
-        // Failure to allocate segment puts determinism at risk, bail out to be safe.
-        // May want configurable behavior here at some point.
-        // If we allow continuing, then consider blacklisting the interval for a while to avoid constant checks.
-        throw new ISE("Could not allocate segment for row with timestamp[%s]", row.getTimestamp());
-      }
-
-      // remember to call incrementNextOffsets before using any other method
-      // which relies on updated value of nextPartitionOffsets
-      if (driver.numRowsInSegment(identifier) >= driver.getMaxRowPerSegment()) {
-        maxRowsPerSegmentLimitReached = true;
-      }
-      return identifier;
-    }
-
-    void incrementNextOffsets(ConsumerRecord<byte[], byte[]> record)
-    {
-      // we do not automatically increment next offset with add call
-      // as in case record cannot be parsed add method will not be called
-      // however we still need to increment next offsets irrespective
-
-      // update local nextOffset to be used by committer at some point
-      nextPartitionOffsets.put(record.partition(), nextPartitionOffsets.get(record.partition()) + 1);
-
-      if (nextPartitionOffsets.get(record.partition()).equals(endOffsets.get(record.partition()))) {
-        logInfo("[Increment-Offsets] read partition [%d]", record.partition());
-        // done with this partition, remove it from end offsets
-        assignment.remove(record.partition());
-        if (assignment.isEmpty()) {
-          logInfo("[Increment-Offsets] read all partitions");
-          Preconditions.checkState(
-              driverStatus == DriverStatus.OPEN,
-              "WTH?! Cannot change driver status to [%s] from [%s]",
-              DriverStatus.COMPLETE,
-              driverStatus
-          );
-          driverStatus = DriverStatus.COMPLETE;
-        }
-      }
-    }
-
-    boolean isCheckPointingRequired()
-    {
-      return maxRowsPerSegmentLimitReached && !checkPointed;
-    }
-
-    boolean isComplete()
-    {
-      return driverStatus == DriverStatus.COMPLETE;
-    }
-
-    void setEndOffsets(Map<Integer, Long> offsets)
-    {
-      // sanity check again with driver's local offsets
-      if (!endOffsets.keySet().containsAll(offsets.keySet())) {
-        throw new IAE(
-            "Got request to set some offsets not handled by driver [%d], handling [%s] partitions",
-            endOffsets.keySet()
-        );
-      }
-      for (Map.Entry<Integer, Long> entry : offsets.entrySet()) {
-        if (entry.getValue().compareTo(nextPartitionOffsets.get(entry.getKey())) < 0) {
-          throw new IAE(
-              "End offset must be >= current offset for driver [%d] for partition [%s] (current: %s)",
-              driverIndex,
-              entry.getKey(),
-              nextPartitionOffsets.get(entry.getKey())
-          );
-        }
-      }
-      endOffsets.putAll(offsets);
-      // check if any or all partitions have been consumed
-      for (Map.Entry<Integer, Long> entry : endOffsets.entrySet()) {
-        if (entry.getValue().compareTo(nextPartitionOffsets.get(entry.getKey())) == 0) {
-          assignment.remove(entry.getKey());
-          logInfo("[Set-End-Offsets] read partition [%d]", entry.getKey());
-        }
-      }
-      logInfo("endOffsets changed to [%s]", endOffsets);
-      if (assignment.isEmpty()) {
-        logInfo("[Set-End-Offsets] read all partitions");
-        Preconditions.checkState(
-            driverStatus == DriverStatus.OPEN,
-            "WTH?! Cannot change driver status to [%s] from [%s]",
-            DriverStatus.COMPLETE,
-            driverStatus
-        );
-        driverStatus = DriverStatus.COMPLETE;
-      }
-      checkPointed = true;
-      logInfo("check-pointed");
-    }
-
-    SegmentsAndMetadata finish(final TaskToolbox toolbox, final boolean isUseTransaction)
-        throws InterruptedException
-    {
-      try {
-        final TransactionalSegmentPublisher publisher = new TransactionalSegmentPublisher()
-        {
-          @Override
-          public boolean publishSegments(Set<DataSegment> segments, Object commitMetadata) throws IOException
-          {
-            final KafkaPartitions finalPartitions = toolbox.getObjectMapper().convertValue(
-                ((Map) commitMetadata).get(METADATA_NEXT_PARTITIONS),
-                KafkaPartitions.class
-            );
-
-            // Sanity check, we should only be publishing things that match our desired end state.
-            if (!endOffsets.equals(finalPartitions.getPartitionOffsetMap())) {
-              throw new ISE("WTF?! Driver attempted to publish invalid metadata[%s].", commitMetadata);
-            }
-
-            final SegmentTransactionalInsertAction action;
-
-            if (isUseTransaction) {
-              action = new SegmentTransactionalInsertAction(
-                  segments,
-                  new KafkaDataSourceMetadata(new KafkaPartitions(topic, startOffsets)),
-                  new KafkaDataSourceMetadata(finalPartitions)
-              );
-            } else {
-              action = new SegmentTransactionalInsertAction(segments, null, null);
-            }
-
-            logInfo("Publishing with isTransaction[%s].", isUseTransaction);
-
-            return toolbox.getTaskActionClient().submit(action).isSuccess();
-          }
-        };
-
-        return driver.finish(publisher, committerSupplier.get());
-      }
-      catch (InterruptedException | RejectedExecutionException e) {
-        logError("Interrupted in finish");
-        throw e;
-      }
-    }
-
-    void waitForHandOff() throws InterruptedException
-    {
-      driver.waitForHandOff();
-    }
-
-    Object persist() throws InterruptedException
-    {
-      return driver.persist(committerSupplier.get());
-    }
-
-    public DriverMetadata getMetadata()
-    {
-      return new DriverMetadata(
-          topic,
-          startOffsets,
-          endOffsets,
-          driverIndex,
-          sequenceName,
-          last,
-          checkPointed,
-          maxRowsPerSegmentLimitReached
-      );
-    }
-
-    @Override
-    public void close() throws IOException
-    {
-      if (driverStatus == DriverStatus.CLOSED) {
-        logWarn("already closed");
-      } else {
-        synchronized (this) {
-          if (driverStatus == DriverStatus.CLOSED) {
-            logWarn("already closed");
-            return;
-          }
-          driverStatus = DriverStatus.CLOSED;
-        }
-        driver.getAppenderator().close();
-        driver.close();
-      }
-    }
-
-    private void logInfo(String msg, Object... formatArgs)
-    {
-      log.info("Driver [%d], status [%s]: [%s]", driverIndex, driverStatus, String.format(msg, formatArgs));
-    }
-
-    private void logWarn(String msg, Object... formatArgs)
-    {
-      log.warn("Driver [%d], status [%s]: [%s]", driverIndex, driverStatus, String.format(msg, formatArgs));
-    }
-
-    private void logError(String msg, Object... formatArgs)
-    {
-      log.error("Driver [%d], status [%s]: [%s]", driverIndex, driverStatus, String.format(msg, formatArgs));
-    }
-
-    @Override
-    public String toString()
-    {
-      return "DriverHolder{" +
-             "driverStatus=" + driverStatus +
-             ", driver=" + driver +
-             ", assignment=" + assignment +
-             ", committerSupplier=" + committerSupplier +
-             ", nextPartitionOffsets=" + nextPartitionOffsets +
-             ", topic='" + topic + '\'' +
-             ", startOffsets=" + startOffsets +
-             ", endOffsets=" + endOffsets +
-             ", sequenceName='" + sequenceName + '\'' +
-             ", driverIndex=" + driverIndex +
-             ", last=" + last +
-             ", checkPointed=" + checkPointed +
-             ", maxRowsPerSegmentLimitReached=" + maxRowsPerSegmentLimitReached +
-             '}';
-    }
-
-    static DriverHolder getNextDriverHolder(
-        KafkaIndexTask taskContext,
-        Map<Integer, Long> startOffsets,
-        Map<Integer, Long> endOffsets,
-        String baseSequenceName,
-        FireDepartmentMetrics fireDepartmentMetrics,
-        TaskToolbox taskToolbox
-    )
-    {
-      return createDriverHolder(
-          taskContext,
-          startOffsets,
-          endOffsets,
-          taskContext.nextDriverIndex++,
-          baseSequenceName,
-          fireDepartmentMetrics,
-          taskToolbox,
-          false,
-          false,
-          false
-      );
-    }
-
-    static DriverHolder getNextDriverHolder(
-        KafkaIndexTask taskContext,
-        Map<Integer, Long> startOffsets,
-        Map<Integer, Long> endOffsets,
-        String baseSequenceName,
-        FireDepartmentMetrics fireDepartmentMetrics,
-        TaskToolbox taskToolbox,
-        boolean checkPointed
-    )
-    {
-      return createDriverHolder(
-          taskContext,
-          startOffsets,
-          endOffsets,
-          taskContext.nextDriverIndex++,
-          baseSequenceName,
-          fireDepartmentMetrics,
-          taskToolbox,
-          false,
-          checkPointed,
-          false
-      );
-    }
-
-    static DriverHolder createDriverHolder(
-        KafkaIndexTask taskContext,
-        Map<Integer, Long> startOffsets,
-        Map<Integer, Long> endOffsets,
-        int basePersistDirectoryIndex,
-        String baseSequenceName,
-        FireDepartmentMetrics fireDepartmentMetrics,
-        TaskToolbox taskToolbox,
-        boolean last,
-        boolean checkPointed,
-        boolean maxRowsPerSegmentLimitReached
-    )
-    {
-      final File basePersistDir = new File(
-          taskToolbox.getTaskWorkDir(),
-          String.format("%s%s", "persist", basePersistDirectoryIndex > 0 ? basePersistDirectoryIndex : "")
-      );
-      DriverHolder holder = new DriverHolder(
-          taskContext.getIOConfig().getStartPartitions().getTopic(),
-          taskContext.newDriver(taskContext.newAppenderator(
-              fireDepartmentMetrics,
-              taskToolbox,
-              basePersistDir
-          ), taskToolbox, fireDepartmentMetrics),
-          startOffsets,
-          endOffsets,
-          basePersistDirectoryIndex,
-          String.format("%s_%d", baseSequenceName, basePersistDirectoryIndex),
-          last,
-          checkPointed,
-          maxRowsPerSegmentLimitReached
-      );
-      log.info("Created new Driver with metadata [%s]", holder.getMetadata());
-      return holder;
-    }
-
-    public static class DriverMetadata implements Comparable<DriverMetadata>
-    {
-      private final String topic;
-      private final Map<Integer, Long> startOffsets;
-      private final Map<Integer, Long> endOffsets;
-      private final int driverIndex;
-      private final String sequenceName;
-      private final boolean last;
-      private final boolean checkPointed;
-      private final boolean maxRowsPerSegmentLimitReached;
-
-      @JsonCreator
-      public DriverMetadata(
-          @JsonProperty("topic") String topic,
-          @JsonProperty("startOffsets") Map<Integer, Long> startOffsets,
-          @JsonProperty("endOffsets") Map<Integer, Long> endOffsets,
-          @JsonProperty("driverIndex") int driverIndex,
-          @JsonProperty("sequenceName") String sequenceName,
-          @JsonProperty("last") boolean last,
-          @JsonProperty("checkPointed") boolean checkPointed,
-          @JsonProperty("maxRowsPerSegmentLimitReached") boolean maxRowsPerSegmentLimitReached
-      )
-      {
-        this.topic = topic;
-        this.startOffsets = startOffsets;
-        this.endOffsets = endOffsets;
-        this.driverIndex = driverIndex;
-        this.sequenceName = sequenceName;
-        this.last = last;
-        this.checkPointed = checkPointed;
-        this.maxRowsPerSegmentLimitReached = maxRowsPerSegmentLimitReached;
-      }
-
-      @JsonProperty
-      public int getDriverIndex()
-      {
-        return driverIndex;
-      }
-
-      @JsonProperty
-      public String getSequenceName()
-      {
-        return sequenceName;
-      }
-
-      @JsonProperty
-      public Map<Integer, Long> getEndOffsets()
-      {
-        return endOffsets;
-      }
-
-      @JsonProperty
-      public Map<Integer, Long> getStartOffsets()
-      {
-        return startOffsets;
-      }
-
-      @JsonProperty
-      public String getTopic()
-      {
-        return topic;
-      }
-
-      @JsonProperty
-      public boolean isLast()
-      {
-        return last;
-      }
-
-      @JsonProperty
-      public boolean isCheckPointed()
-      {
-        return checkPointed;
-      }
-
-      @JsonProperty
-      public boolean isMaxRowsPerSegmentLimitReached()
-      {
-        return maxRowsPerSegmentLimitReached;
-      }
-
-      @Override
-      public String toString()
-      {
-        return "DriverMetadata{" +
-               "topic='" + topic + '\'' +
-               ", startOffsets=" + startOffsets +
-               ", endOffsets=" + endOffsets +
-               ", driverIndex=" + driverIndex +
-               ", sequenceName='" + sequenceName + '\'' +
-               ", last=" + last +
-               ", checkPointed=" + checkPointed +
-               ", maxRowsPerSegmentLimitReached=" + maxRowsPerSegmentLimitReached +
-               '}';
-      }
-
-      @Override
-      public int compareTo(DriverMetadata o)
-      {
-        return driverIndex - o.driverIndex;
-      }
-    }
-  }
-
-  private static class SentinelDriverHolder extends DriverHolder
-  {
-    final CountDownLatch persistLatch;
-    final CountDownLatch handOffLatch;
-
-    private SentinelDriverHolder(
-        CountDownLatch persistLatch,
-        CountDownLatch handOffLatch
-    )
-    {
-      super(null, null, ImmutableMap.<Integer, Long>of(), ImmutableMap.<Integer, Long>of(), -1, null, true, true, true);
-      super.driverStatus = DriverStatus.COMPLETE;
-      this.persistLatch = persistLatch;
-      this.handOffLatch = handOffLatch;
-    }
-
-    @Override
-    Object persist() throws InterruptedException
-    {
-      persistLatch.countDown();
-      return null;
-    }
-
-    @Override
-    SegmentsAndMetadata finish(TaskToolbox toolbox, boolean isUseTransaction) throws InterruptedException
-    {
-      // Do nothing
-      return null;
-    }
-
-    @Override
-    public void close() throws IOException
-    {
-      handOffLatch.countDown();
     }
   }
 }
